Calculate Tree Biomass for National Forest Inventory Data

biomass_nfi() is a function that estimates volume, aboveground biomass, biomass, carbon storage and carbon dioxide storage. It can provide summaries for individual plots, the entire study area, or specific groups within the study area using parameters byplot, plotgrp and treegrp. When calculating biomass at the individual trees and plots level, Users have flexibility in specifying data inclusion criteria and analysis levels using parameters clusterplot, largetreearea, stockedland, and talltree. These parameters determine whether to treat cluster plots as single plots, to include large tree survey plots, and to focus only on Stocked land and tall trees. Users can also choose the criteria for post-stratification using the strat parameter.

Usage

biomass_nfi(
  data,
  byplot = FALSE,
  plotgrp = NULL,
  treegrp = NULL,
  strat = "FORTYP_SUB",
  clusterplot = FALSE,
  largetreearea = TRUE,
  stockedland = TRUE,
  talltree = TRUE
)

Arguments

data

: A list generated by read_nfi that contains 'plot' and 'tree' data frames.

byplot

: A logical flag (default FALSE); if TRUE, calculates statistics for each plot separately. If FALSE, calculates for the entire dataset or groups specified by plotgrp and treegrp.

plotgrp

: A character vector; variables from 'plot' tables for grouping. Use c() to combine multiple variables.

treegrp

: A character vector; variables from 'tree' tables for grouping. Use c() to combine multiple variables.

strat

: A character vector; the variable used for post-stratification. In the National Forest Inventory of Korea, it is typically used by forest type.

clusterplot

: A logical flag (default FALSE); if TRUE, treats each cluster plot as a single unit. If FALSE, calculates for each subplot separately.

largetreearea

: A logical flag (default TRUE); if TRUE, includes large tree survey plots in the analysis. If FALSE, only uses standard tree plots.

stockedland

: A logical flag (default TRUE); if TRUE, includes only stocked land. If FALSE, includes all land types.

talltree

: A logical flag (default TRUE); if TRUE, includes only tall trees. If FALSE, includes both trees and shrubs.

Value

A data.frame that includes biomass estimates. The structure depends on the input parameters:

If byplot = TRUE, each row represents a plot.

If byplot = FALSE, each row represents the entire dataset or a group specified by plotgrp and treegrp

Details

This function calculates biomass using methodologies employed for national statistics (mean, variance, standard error, relative standard error):

Applies national carbon emission factors to calculate biomass at the individual tree level.

Estimates biomass per hectare at the cluster or subplot level, with options to include only basic survey trees or both basic and large tree survey trees.

Uses the Double Sampling for Post-stratification (DSS) method to derive annual statistics.

Applies the Weighted Moving Average (WMA) method to integrate annual statistics from the 20% of plots surveyed each year into a single time point.

Note

Biomass calculation involves dividing the data into groups based on plotgrp and then applying post-stratification to each group. As a result, if the data for each group is not sufficiently large, the relative standard error (RSE) may be high. It is important to check the RSE and other statistical measures in the biomass results.

References

Son, Y., Kim, R., Lee, K., Pyo, J., Kim, S., Hwang, J., Lee, S., & Park, H. (2014). Carbon emission factors and biomass allometric equations by species in Korea. Korea Forest Research Institute. Yim, J., Moon, G., Lee, M., Kang, J., Won, M., Ahn, E., & Jeon, J. (2021). 2020 Forest inventory of Korea. Korea Forest Research Institute.

Examples

data("nfi_sejong")

# Basic usage
biomass <- biomass_nfi(nfi_sejong)

if (FALSE) { # \dontrun{
# Calculate biomass by administrative district
district_biomass <- biomass_nfi(nfi_sejong, plotgrp = "SGG")

# Calculate biomass for each plot
plot_biomass <- biomass_nfi(nfi_sejong, byplot = TRUE)
} # }