Calculates the importance values for National Forest Inventory Data

iv_nfi() is a function that calculates the importance values of tree species based on frequency, density and coverage. It can provide estimates for the entire study area, or specific groups within it using the plotgrp parameter. It uses the importancevalue function from the BiodiversityR package for core calculations. Users have flexibility in specifying data inclusion criteria and analysis levels using parameters frequency, clusterplot, largetreearea, stockedland, and talltree. These parameters determine whether to include frequency in importance calculations, to treat cluster plots as single plots, to include large tree survey plots, and to focus only on Stocked land and tall trees.

Usage

iv_nfi(
  data,
  sp = "SP",
  plotgrp = NULL,
  frequency = TRUE,
  clusterplot = FALSE,
  largetreearea = FALSE,
  stockedland = TRUE,
  talltree = TRUE
)

Arguments

data

: A list generated by read_nfi that contains 'plot' and 'tree' data frames.

sp

:A character vector; the column name of species information (e.g., "SP" for species, "GENUS" for genus-level analysis).

plotgrp

: A character vector; specifies variables from 'plot' table to use for grouping. Use c() to combine multiple variables.

frequency

: A logical flag (default TRUE); if TRUE, includes frequency in importance value calculations.

clusterplot

: A logical flag (default FALSE); if TRUE, treats each cluster plot as a single unit. If FALSE, calculates for each subplot separately.

largetreearea

: A logical flag (default FALSE); if TRUE, includes large tree survey plots in the analysis. If FALSE, only uses standard tree plots.

stockedland

: A logical flag (default TRUE); if TRUE, includes only stocked land. If FALSE, includes all land types.

talltree

: A logical flag (default TRUE); if TRUE, includes only tall trees. If FALSE, includes both trees and shrubs.

Value

A data.frame that includes importance value for tree species. Each row represents a combination of a tree species and groups specified by plotgrp and treegrp.

Details

The importance value (ranging from 0 to 100) is calculated as the mean of:

Relative frequency: (number of plots where a species is observed / total survey plots) * 100

Relative density: (total number of individuals of a species / sum of all species' densities) * 100

Relative coverage: (total basal area of a species / sum of all species' basal area) * 100

Note

Consider calculating importance by genus rather than species due to potential incompleteness in species classification.

Since the frequencies of each species may be identical across the nation, it may be desirable to exclude frequency from the importance calculation.

References

Curtis, J. T. & McIntosh, R. P. (1951). An upland forest continuum in the prairie-forest border region of Wisconsin. Ecology, 32(3), 476–496.

See also

Examples

data("nfi_sejong")
# Calculate importance values without frequency
importance <- iv_nfi(nfi_sejong, sp = "SP", frequency = FALSE)

if (FALSE) { # \dontrun{
# Calculate importance values using genus
genus_importance <- iv_nfi(nfi_sejong, sp = "GENUS")
} # }